improve treatment of split strings, add cut label
In some rare circumstances a graphic string has to be split
in 2 smaller pieces which will be printed on different lines.
A new option `--split-bytes SPLI-MIN` gives finer control
under which condition a piece shall be printed.

Furthermore both pieces are labelled with a flag symbol ⚑ (U+2691)
to indicate that they had been cut.